requirements.txt
pandas==2.1.0
scikit-learn==1.3.0
SQLAlchemy==2.0.20
gunicorn==20.1.0
psycopg2-binary==2.9.7
Flask==2.3.3
spanlp==1.1.0
Flask-Cors==4.0.0
# sentence-transformers==2.2.2
pymongo==4.5.0
scipy
numpy
requests
python-docx


